Move drawing for panel simulation to separate file
